@@ -1,6 +1,6 @@
 {
   "name": "crafatar",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "private": true,
   "description": "A blazing fast API for Minecraft faces!",
   "contributors": [
@@ -31,23 +31,23 @@
     "test-travis": "istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage"
   },
   "engines": {
-    "node": "8.9.4"
+    "node": "10.19.0"
   },
   "dependencies": {
-    "canvas": "^1.6.2",
-    "crc": "^3.5.0",
-    "ejs": "^2.5.2",
-    "pajk-lwip": "^0.2.0",
-    "mime": "^2.2.0",
+    "canvas": "^2.6.1",
+    "crc": "^3.8.0",
+    "ejs": "^3.0.1",
+    "@randy.tarampi/lwip": "^1.1.0",
+    "mime": "^2.4.4",
     "node-df": "crafatar/node-df",
-    "redis": "^2.8.0",
-    "request": "^2.83.0",
+    "redis": "^3.0.2",
+    "request": "^2.88.2",
     "toobusy-js": "^0.5.1"
   },
   "devDependencies": {
-    "coveralls": "^3.0.0",
+    "coveralls": "^3.0.11",
     "istanbul": "^0.4.5",
-    "mocha": "^5.0.1",
+    "mocha": "^7.1.1",
     "mocha-lcov-reporter": "^1.3.0"
   }
 }
