@@ -91,7 +91,7 @@ exp.get_uuid_info = function(profile, type, callback) {
   var properties = Object.get(profile, "properties") || [];
   properties.forEach(function(prop) {
     if (prop.name === "textures") {
-      var json = new Buffer(prop.value, "base64").toString();
+      var json = new Buffer.from(prop.value, "base64").toString();
       profile = JSON.parse(json);
     }
   });
