@@ -2,7 +2,7 @@ var helpers = require("../helpers");
 var skins = require("../skins");
 var cache = require("../cache");
 var path = require("path");
-var lwip = require("pajk-lwip");
+var lwip = require("@randy.tarampi/lwip");
 var url = require("url");
 
 // handle the appropriate 'default=' response
