@@ -1,5 +1,5 @@
 var logging = require("./logging");
-var lwip = require("pajk-lwip");
+var lwip = require("@randy.tarampi/lwip");
 var fs = require("fs");
 
 var exp = {};
